dotcom: migrate Cody Gateway rate limit to int64 (#53290)
This PR changes the rate limit quota type from `int32` to
`BigInt`/`int64`.

## Test plan

Manually tested in combination with
https://github.com/sourcegraph/sourcegraph/pull/53289


Update rate limit:


https://github.com/sourcegraph/sourcegraph/assets/2946214/ff468073-9bc4-4413-9e64-4bc40fd5768f

Render data points:

<img width="1029" alt="CleanShot 2023-06-16 at 11 23 28@2x"
src="https://github.com/sourcegraph/sourcegraph/assets/2946214/185fc32e-228c-43f9-8f5f-eba226d34013">


---

~~Stack on https://github.com/sourcegraph/sourcegraph/pull/53289,~~ part
of https://github.com/sourcegraph/sourcegraph/issues/53007

+++ enterprise/cmd/frontend/internal/dotcom/productsubscription/codygateway_dotcom_user.go
@@ -145,7 +145,7 @@ func (r codyUserGatewayAccessResolver) EmbeddingsRateLimit(ctx context.Context) 
 
 	rateLimit := licensing.CodyGatewayRateLimit{
 		AllowedModels:   []string{"openai/text-embedding-ada-002"},
-		Limit:           int32(20 * tokensPerDollar),
+		Limit:           int64(20 * tokensPerDollar),
 		IntervalSeconds: math.MaxInt32,
 	}
 
@@ -195,7 +195,7 @@ func getCompletionsRateLimit(ctx context.Context, db database.DB, userID int32, 
 	}
 	return licensing.CodyGatewayRateLimit{
 		AllowedModels:   allowedModels(scope),
-		Limit:           int32(*limit),
+		Limit:           int64(*limit),
 		IntervalSeconds: 86400, // Daily limit TODO(davejrt)
 	}, source, nil
 }
